requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.12
chromaviz @ git+https://github.com/amikos-tech/chromaviz/@6e67080ef60e5a0e065a3a28686986dddab66b15
clarifai==9.7.1
clarifai-grpc==9.8.3
click==8.1.7
cmake==3.27.5
cohere==4.27
coloredlogs==15.0.1
dataclasses-json==0.6.0
dill==0.3.7
fastapi==0.99.1
fastavro==1.8.2
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.300
langsmith==0.0.40
lit==16.0.6
manifest-ml==0.0.1
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nlpcloud==1.1.44
numexpr==2.8.6
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.16.0
openai==0.28.0
openlm==0.0.5
overrides==7.4.0
packaging==23.1
pandas==2.1.1
posthog==3.0.2
protobuf==4.24.3
pulsar-client==3.3.0
pydantic==1.10.12
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.11
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlitedict==2.1.0
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.2
triton==2.0.0
tritonclient==2.34.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.17.0
